+++ Package.swift
@@ -13,9 +13,15 @@ let package = Package(
             targets: ["OtplessBM"]
         ),
     ],
+    dependencies: [
+        .package(url: "https://github.com/otpless-tech/otpless-event-io-ios.git", from: "1.0.0")
+    ],
     targets: [
         .target(
-            name: "OtplessBM"
+            name: "OtplessBM",
+            dependencies: [
+                .product(name: "OtplessEventIO", package: "otpless-event-io-ios")
+            ]
         ),
     ],
     swiftLanguageVersions: [.v5]